remote_service/requirements.txt
fastapi
pillow
pymupdf==1.25.5
gemma==3.0.2
jax[cuda12]
loguru
ngrok
uvicorn
python-dotenv